delegating the api functions to a designated file
